# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API documentation example",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://test.naufalrafif.com/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://staging.example.com/api",
            "description": "Staging Server"
        },
        {
            "url": "https://api.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/data-master/district": {
            "get": {
                "tags": [
                    "Data Master - District"
                ],
                "summary": "Get all district",
                "operationId": "4b762d999df481313aeddaad3d88f2b4",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by district name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit data per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page data",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "regency_id",
                        "in": "query",
                        "description": "Regency id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "meta": {
                                            "properties": {
                                                "total_data": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total_pages": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "alt_name": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "latitude": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "longitude": {
                                                        "type": "string",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Data Master - District"
                ],
                "summary": "Create a district",
                "description": "Create a district",
                "operationId": "58a71187614babdd3b6499d36605ac31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jakarta Pusat"
                                    },
                                    "alt_name": {
                                        "type": "string",
                                        "example": "Jakarta Pusat"
                                    },
                                    "latitude": {
                                        "type": "string",
                                        "example": "-6.21462"
                                    },
                                    "longitude": {
                                        "type": "string",
                                        "example": "106.84513"
                                    },
                                    "regency_id": {
                                        "type": "string",
                                        "example": "regency-1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "alt_name": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "latitude": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "longitude": {
                                                    "type": "string",
                                                    "example": ""
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}